+++ customers.go
@@ -38,8 +38,8 @@ type Customer struct {
 	// Email: Email address for this customer.
 	Email string `json:"email,omitempty"`
 	/*
-	   MarketingConsent: Whether this customer opted into marketing from you.
-	   Set to `true` by Paddle where a customer checks the marketing consent box when using Paddle Checkout; `false` otherwise.
+	   MarketingConsent: Whether this customer opted into marketing from you. `false` unless customers check the marketing consent box
+	   when using Paddle Checkout. Set automatically by Paddle.
 	*/
 	MarketingConsent bool `json:"marketing_consent,omitempty"`
 	// Status: Whether this entity can be used in Paddle.
@@ -152,7 +152,7 @@ type UpdateCustomerRequest struct {
 	Status *PatchField[Status] `json:"status,omitempty"`
 	// CustomData: Your own structured key-value data.
 	CustomData *PatchField[CustomData] `json:"custom_data,omitempty"`
-	// Locale: Valid IETF BCP 47 short form locale tag. If omitted, defaults to `en`.
+	// Locale: Valid IETF BCP 47 short form locale tag.
 	Locale *PatchField[string] `json:"locale,omitempty"`
 }