+++ pkg/paddlenotification/adjustments.go
@@ -16,7 +16,7 @@ type AdjustmentUpdated struct {
 	Data AdjustmentNotification `json:"data"`
 }
 
-// Action: How this adjustment impacts the related transaction. `refund` adjustments must be approved by Paddle, and are created with the status `pending_approval`..
+// Action: How this adjustment impacts the related transaction..
 type Action string
 
 const (
@@ -31,9 +31,9 @@ const (
 /*
 AdjustmentStatus: Status of this adjustment. Set automatically by Paddle.
 
-`refund` adjustments must be approved by Paddle, and are created with the status `pending_approval`
-until they move to `approved` or `rejected` on review. Other kinds of adjustment do not need approval,
-so are created with the status `approved`..
+Most refunds for live accounts are created with the status of `pending_approval` until reviewed by Paddle, but some are automatically approved. For sandbox accounts, Paddle automatically approves refunds every ten minutes.
+
+Credit adjustments don't require approval from Paddle, so they're created as `approved`..
 */
 type AdjustmentStatus string
 
@@ -45,7 +45,7 @@ const (
 )
 
 /*
-AdjustmentType: Type of adjustment for this transaction item. `tax` and `proration` are automatically created by Paddle.
+AdjustmentType: Type of adjustment for this transaction item. `tax` adjustments are automatically created by Paddle.
 Include `amount` when creating a `partial` adjustment..
 */
 type AdjustmentType string
@@ -74,16 +74,13 @@ type AdjustmentItem struct {
 	// ItemID: Paddle ID for the transaction item that this adjustment item relates to, prefixed with `txnitm_`.
 	ItemID string `json:"item_id,omitempty"`
 	/*
-	   Type: Type of adjustment for this transaction item. `tax` and `proration` are automatically created by Paddle.
+	   Type: Type of adjustment for this transaction item. `tax` adjustments are automatically created by Paddle.
 	   Include `amount` when creating a `partial` adjustment.
 	*/
 	Type AdjustmentType `json:"type,omitempty"`
 	// Amount: Amount adjusted for this transaction item. Required when adjustment type is `partial`.
 	Amount *string `json:"amount,omitempty"`
-	/*
-	   Proration: How proration was calculated for this adjustment item. Populated when an adjustment type is `proration`.
-	   Set automatically by Paddle.
-	*/
+	// Proration: How proration was calculated for this adjustment item.
 	Proration *Proration `json:"proration,omitempty"`
 	// Totals: Breakdown of the total for an adjustment item.
 	Totals AdjustmentItemTotals `json:"totals,omitempty"`
@@ -132,9 +129,9 @@ type PayoutTotalsAdjustment struct {
 type AdjustmentNotification struct {
 	// ID: Unique Paddle ID for this adjustment entity, prefixed with `adj_`.
 	ID string `json:"id,omitempty"`
-	// Action: How this adjustment impacts the related transaction. `refund` adjustments must be approved by Paddle, and are created with the status `pending_approval`.
+	// Action: How this adjustment impacts the related transaction.
 	Action Action `json:"action,omitempty"`
-	// TransactionID: Paddle ID for the transaction related to this adjustment, prefixed with `txn_`.
+	// TransactionID: Paddle ID of the transaction that this adjustment is for, prefixed with `txn_`.
 	TransactionID string `json:"transaction_id,omitempty"`
 	/*
 	   SubscriptionID: Paddle ID for the subscription related to this adjustment, prefixed with `sub_`.
@@ -146,15 +143,14 @@ type AdjustmentNotification struct {
 	   Set automatically by Paddle based on the `customer_id` of the related transaction.
 	*/
 	CustomerID string `json:"customer_id,omitempty"`
-	// Reason: Why this adjustment was created. Appears in the Paddle Dashboard. Retained for record-keeping purposes.
+	// Reason: Why this adjustment was created. Appears in the Paddle dashboard. Retained for record-keeping purposes.
 	Reason string `json:"reason,omitempty"`
 	/*
 	   CreditAppliedToBalance: Whether this adjustment was applied to the related customer's credit balance. Only returned for `credit` adjustments.
 
-	   `false` when the related transaction `collection_mode` is `manual` and its `status` is `billed`. The adjustment is used
-	   to reduce the `balance` due on the transaction.
+	   `false` where the related transaction is `billed`. The adjustment reduces the amount due on the transaction.
 
-	   `true` for automatically-collected transactions and `completed` manually-collected transactions.
+	   `true` where the related transaction is `completed`. The amount is added the customer's credit balance and used to pay future transactions.
 	*/
 	CreditAppliedToBalance *bool `json:"credit_applied_to_balance,omitempty"`
 	// CurrencyCode: Three-letter ISO 4217 currency code for this adjustment. Set automatically by Paddle based on the `currency_code` of the related transaction.
@@ -162,9 +158,9 @@ type AdjustmentNotification struct {
 	/*
 	   Status: Status of this adjustment. Set automatically by Paddle.
 
-	   `refund` adjustments must be approved by Paddle, and are created with the status `pending_approval`
-	   until they move to `approved` or `rejected` on review. Other kinds of adjustment do not need approval,
-	   so are created with the status `approved`.
+	   Most refunds for live accounts are created with the status of `pending_approval` until reviewed by Paddle, but some are automatically approved. For sandbox accounts, Paddle automatically approves refunds every ten minutes.
+
+	   Credit adjustments don't require approval from Paddle, so they're created as `approved`.
 	*/
 	Status AdjustmentStatus `json:"status,omitempty"`
 	// Items: List of items on this adjustment.