+++ pkg/paddlenotification/discounts.go
@@ -33,7 +33,7 @@ const (
 	DiscountStatusUsed     DiscountStatus = "used"
 )
 
-// Type: Type of discount..
+// Type: Type of discount. Determines how this discount impacts the transaction total..
 type Type string
 
 const (
@@ -50,17 +50,17 @@ type DiscountNotification struct {
 	Status DiscountStatus `json:"status,omitempty"`
 	// Description: Short description for this discount for your reference. Not shown to customers.
 	Description string `json:"description,omitempty"`
-	// EnabledForCheckout: Whether this discount can be applied by a customer at checkout.
+	// EnabledForCheckout: Whether this discount can be applied by customers at checkout.
 	EnabledForCheckout bool `json:"enabled_for_checkout,omitempty"`
-	// Code: Unique code that customers can use to apply this discount at checkout. Use letters and numbers only, up to 16 characters. Paddle generates a random 10-character code if a code is not provided and `enabled_for_checkout` is `true`.
+	// Code: Unique code that customers can use to apply this discount at checkout.
 	Code *string `json:"code,omitempty"`
-	// Type: Type of discount.
+	// Type: Type of discount. Determines how this discount impacts the transaction total.
 	Type Type `json:"type,omitempty"`
 	// Amount: Amount to discount by. For `percentage` discounts, must be an amount between `0.01` and `100`. For `flat` and `flat_per_seat` discounts, amount in the lowest denomination for a currency.
 	Amount string `json:"amount,omitempty"`
 	// CurrencyCode: Supported three-letter ISO 4217 currency code. Required where discount type is `flat` or `flat_per_seat`.
 	CurrencyCode *CurrencyCode `json:"currency_code,omitempty"`
-	// Recur: Whether this discount applies for multiple billing periods.
+	// Recur: Whether this discount applies for multiple subscription billing periods.
 	Recur bool `json:"recur,omitempty"`
 	// MaximumRecurringIntervals: Amount of subscription billing periods that this discount recurs for. Requires `recur`. `null` if this discount recurs forever.
 	MaximumRecurringIntervals *int `json:"maximum_recurring_intervals,omitempty"`